// frontend/src/api/models/CreateGAD7TestDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime'
/**
 *
 * @export
 * @interface CreateGAD7TestDTO
 */
export interface CreateGAD7TestDTO {
  /**
   *
   * @type {string}
   * @memberof CreateGAD7TestDTO
   */
  patientId?: string
  /**
   *
   * @type {string}
   * @memberof CreateGAD7TestDTO
   */
  sessionId?: string
  /**
   *
   * @type {number}
   * @memberof CreateGAD7TestDTO
   */
  question1?: number
  /**
   *
   * @type {number}
   * @memberof CreateGAD7TestDTO
   */
  question2?: number
  /**
   *
   * @type {number}
   * @memberof CreateGAD7TestDTO
   */
  question3?: number
  /**
   *
   * @type {number}
   * @memberof CreateGAD7TestDTO
   */
  question4?: number
  /**
   *
   * @type {number}
   * @memberof CreateGAD7TestDTO
   */
  question5?: number
  /**
   *
   * @type {number}
   * @memberof CreateGAD7TestDTO
   */
  question6?: number
  /**
   *
   * @type {number}
   * @memberof CreateGAD7TestDTO
   */
  question7?: number
}

/**
 * Check if a given object implements the CreateGAD7TestDTO interface.
 */
export function instanceOfCreateGAD7TestDTO(value: object): value is CreateGAD7TestDTO {
  return true
}

export function CreateGAD7TestDTOFromJSON(json: any): CreateGAD7TestDTO {
  return CreateGAD7TestDTOFromJSONTyped(json, false)
}

export function CreateGAD7TestDTOFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean
): CreateGAD7TestDTO {
  if (json == null) {
    return json
  }
  return {
    patientId: json['patientId'] == null ? undefined : json['patientId'],
    sessionId: json['sessionId'] == null ? undefined : json['sessionId'],
    question1: json['question1'] == null ? undefined : json['question1'],
    question2: json['question2'] == null ? undefined : json['question2'],
    question3: json['question3'] == null ? undefined : json['question3'],
    question4: json['question4'] == null ? undefined : json['question4'],
    question5: json['question5'] == null ? undefined : json['question5'],
    question6: json['question6'] == null ? undefined : json['question6'],
    question7: json['question7'] == null ? undefined : json['question7'],
  }
}

export function CreateGAD7TestDTOToJSON(json: any): CreateGAD7TestDTO {
  return CreateGAD7TestDTOToJSONTyped(json, false)
}

export function CreateGAD7TestDTOToJSONTyped(
  value?: CreateGAD7TestDTO | null,
  ignoreDiscriminator: boolean = false
): any {
  if (value == null) {
    return value
  }

  return {
    patientId: value['patientId'],
    sessionId: value['sessionId'],
    question1: value['question1'],
    question2: value['question2'],
    question3: value['question3'],
    question4: value['question4'],
    question5: value['question5'],
    question6: value['question6'],
    question7: value['question7'],
  }
}